refactor: 将函数放置在 utils 中
